Add typings
Should be working
