+++ index.js
@@ -1,7 +1,7 @@
-const User = require("./User.js");
-const Media = require("./Media.js");
-const Revoice = require("./Revoice.js");
-const Signaling = require("./Signaling.js");
+const User = require("./src/User.js");
+const Media = require("./src/Media.js");
+const Revoice = require("./src/Revoice.js");
+const Signaling = require("./src/Signaling.js");
 
 module.exports = {
   ...Media, Revoice, Signaling, User
