+++ tsconfig.json
@@ -1,0 +1,27 @@
+{
+    // Change this to match your project
+    "include": [
+        "src/**/*"
+    ],
+    "exclude": ["/node_modules/*"],
+    "compilerOptions": {
+        "target": "ES6",
+        // Tells TypeScript to read JS files, as
+        // normally they are ignored as source files
+        "allowJs": true,
+        // Generate d.ts files
+        "declaration": true,
+        // This compiler run should
+        // only output d.ts files
+        "emitDeclarationOnly": true,
+        // Types should go into this directory.
+        // Removing this would place the .d.ts files
+        // next to the .js files
+        "outDir": "typings",
+        // go to js file when using IDE functions like
+        // "Go to Definition" in VSCode
+        "declarationMap": true,
+        "moduleResolution":"Node",
+        "allowSyntheticDefaultImports": true
+    }
+}
